+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export * from './array'
-export * from './composite-matrix'
+export * from './block-matrix'
 export * from './math'
 export * from './matrix'
 export * from './misc'